+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition/Schema.hs
@@ -40,10 +40,10 @@ module Squeal.PostgreSQL.Definition.Schema
 
 import GHC.TypeLits
 
-import Squeal.PostgreSQL.Alias
+import Squeal.PostgreSQL.Type.Alias
 import Squeal.PostgreSQL.Definition
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
