+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition/Table.hs
@@ -66,14 +66,14 @@ import GHC.TypeLits
 import qualified Generics.SOP as SOP
 import qualified GHC.Generics as GHC
 
-import Squeal.PostgreSQL.Alias
+import Squeal.PostgreSQL.Type.Alias
 import Squeal.PostgreSQL.Definition
 import Squeal.PostgreSQL.Definition.Constraint
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Type
-import Squeal.PostgreSQL.List
+import Squeal.PostgreSQL.Type.List
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
