+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition/Type.hs
@@ -49,14 +49,14 @@ import Prelude hiding ((.), id)
 
 import qualified Generics.SOP as SOP
 
-import Squeal.PostgreSQL.Alias
+import Squeal.PostgreSQL.Type.Alias
 import Squeal.PostgreSQL.Expression.Logic
 import Squeal.PostgreSQL.Expression.Type
 import Squeal.PostgreSQL.Definition
-import Squeal.PostgreSQL.List
-import Squeal.PostgreSQL.PG
+import Squeal.PostgreSQL.Type.List
+import Squeal.PostgreSQL.Type.PG
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
