+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression.hs
@@ -77,10 +77,10 @@ import Prelude hiding (id, (.))
 
 import qualified GHC.Generics as GHC
 
-import Squeal.PostgreSQL.Alias
-import Squeal.PostgreSQL.List
+import Squeal.PostgreSQL.Type.Alias
+import Squeal.PostgreSQL.Type.List
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
