+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Comparison.hs
@@ -48,7 +48,7 @@ import Data.ByteString
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Logic
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
