+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Json.hs
@@ -79,14 +79,14 @@ module Squeal.PostgreSQL.Expression.Json
 import Data.ByteString (ByteString)
 import GHC.TypeLits
 
-import Squeal.PostgreSQL.Alias
+import Squeal.PostgreSQL.Type.Alias
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Set
 import Squeal.PostgreSQL.Expression.Type
-import Squeal.PostgreSQL.List
+import Squeal.PostgreSQL.Type.List
 import Squeal.PostgreSQL.Query
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 import qualified Generics.SOP as SOP
