+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Null.hs
@@ -32,7 +32,7 @@ module Squeal.PostgreSQL.Expression.Null
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Logic
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
