+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Range.hs
@@ -68,9 +68,9 @@ import qualified Generics.SOP as SOP
 
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Type hiding (bool)
-import Squeal.PostgreSQL.PG
+import Squeal.PostgreSQL.Type.PG
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL (tstzrange, numrange, int4range, now, printSQL)
