+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Window.hs
@@ -59,14 +59,14 @@ import Data.ByteString (ByteString)
 import qualified GHC.Generics as GHC
 import qualified Generics.SOP as SOP
 
-import Squeal.PostgreSQL.Alias
+import Squeal.PostgreSQL.Type.Alias
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Aggregate
 import Squeal.PostgreSQL.Expression.Logic
 import Squeal.PostgreSQL.Expression.Sort
-import Squeal.PostgreSQL.List
+import Squeal.PostgreSQL.Type.List
 import Squeal.PostgreSQL.Render
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 instance Aggregate (WindowArg grp) (WindowFunction grp) where
   countStar = UnsafeWindowFunction "count(*)"
