+++ squeal-postgresql/src/Squeal/PostgreSQL/Manipulation.hs
@@ -67,16 +67,16 @@ import qualified Generics.SOP as SOP
 import qualified Generics.SOP.Record as SOP
 import qualified GHC.Generics as GHC
 
-import Squeal.PostgreSQL.Alias
+import Squeal.PostgreSQL.Type.Alias
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Default
 import Squeal.PostgreSQL.Expression.Inline
 import Squeal.PostgreSQL.Expression.Logic
-import Squeal.PostgreSQL.List
-import Squeal.PostgreSQL.PG
+import Squeal.PostgreSQL.Type.List
+import Squeal.PostgreSQL.Type.PG
 import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Query
-import Squeal.PostgreSQL.Schema
+import Squeal.PostgreSQL.Type.Schema
 
 -- $setup
 -- >>> import Squeal.PostgreSQL
